requirements.txt
# GenMamba-Flow: Robust Generative Steganography
# Python 3.8+

torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.62.0
tensorboard>=2.10.0
opencv-python>=4.5.0
einops>=0.6.0
lpips>=0.1.4
clean-fid>=0.1.0
ftfy
regex
open-clip-torch>=2.20.0
# Optional: for faster Mamba (Linux + CUDA only)
# mamba-ssm>=2.0.0
# causal-conv1d>=1.4.0
